+++ src/datasource/CartDataSource.ts
@@ -1,22 +1,23 @@
-import { baseURL } from "../config/develop-config";
+import { BASE_URL_SOHEE } from "../config/config";
 import { CartEntity } from "../entities/CartEntity";
 import { ShipmentEntity } from "../entities/ShipmentEntity";
 import { ItemSpecialRequest } from "../models/SpecialRequestModel";
 import { httpClient } from "../services/HttpClient";
 
 export class CartDataSource {
-    static getCartByShop(shopId: string, productBrand?: string): Promise<CartEntity> {
+    static getCartByShop(company: string, shopId: string, productBrand?: string): Promise<CartEntity> {
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .get(`${baseURL}/v1/sellcoda/cart`, { params })
+            .get(`${BASE_URL_SOHEE}/v1/sellcoda/cart`, { params })
             .then(res => res.data)
             .catch(error => console.log(error.response.data))
     }
 
-    static addToCartByShopId(shopId: string, itemId: string, quantity: number, payment?: string, useSubsidize?: boolean, productBrand?: string) {
+    static addToCartByShopId(company: string, shopId: string, itemId: string, quantity: number, payment?: string, useSubsidize?: boolean, productBrand?: string) {
         const data = {
             action: "adjust",
             item_id: itemId,
@@ -26,15 +27,16 @@ export class CartDataSource {
         }
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart`, data, { params })
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart`, data, { params })
             .then(res => res.data)
             .catch(error => console.log(error))
     }
 
-    static removeItem(shopId: string, itemId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
+    static removeItem(company: string, shopId: string, itemId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
         const data = {
             action: "remove",
             item_id: itemId,
@@ -43,15 +45,16 @@ export class CartDataSource {
         }
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart`, data, { params })
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart`, data, { params })
             .then(res => res.data)
             .catch(error => console.log(error))
     }
 
-    static calculate(shopId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
+    static calculate(company: string, shopId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
         const data = {
             action: "calculate",
             payment_method: payment,
@@ -59,22 +62,24 @@ export class CartDataSource {
         }
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart`, data, { params })
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart`, data, { params })
             .then(res => res.data)
             .catch(error => console.log(error))
     }
 
-    static calculateSpecialRequest(shopId: string, discounts: ItemSpecialRequest[], remark: string, productBrand?: string) {
+    static calculateSpecialRequest(company: string, shopId: string, discounts: ItemSpecialRequest[], remark: string, productBrand?: string) {
         let arr = { discounts, remark }
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart/calculate_special_request`, arr, { params })
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart/calculate_special_request`, arr, { params })
             .then(res => res.data)
             .catch(error => console.log(error))
     }
@@ -82,12 +87,12 @@ export class CartDataSource {
     static createSpecialRequest(shopId: string, discounts: ItemSpecialRequest[], remark: string): Promise<string> {
         let arr = { discounts, remark }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart/special_request?shopId=${shopId}`, arr)
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart/special_request?shopId=${shopId}`, arr)
             .then(res => res.data.id)
             .catch(error => console.log(error))
     }
 
-    static addSpecialRequest(shopId: string, specialId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
+    static addSpecialRequest(company: string, shopId: string, specialId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
         const data = {
             action: "add_special_request",
             special_request_id: specialId,
@@ -96,15 +101,16 @@ export class CartDataSource {
         }
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart`, data, { params })
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart`, data, { params })
             .then(res => res.data)
             .catch(error => console.log(error))
     }
 
-    static clearSpecialRequest(shopId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
+    static clearSpecialRequest(company: string, shopId: string, payment?: string, useSubsidize?: boolean, productBrand?: string) {
         const data = {
             action: "remove_special_request",
             payment_method: payment,
@@ -112,31 +118,34 @@ export class CartDataSource {
         }
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart`, data, { params })
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart`, data, { params })
             .then(res => res.data)
             .catch(error => console.log(error))
     }
 
-    static clearCart(shopId: string, productBrand?: string) {
+    static clearCart(company: string, shopId: string, productBrand?: string) {
         const data = {
             action: "clear",
         }
         const params = {
             shopId: shopId,
+            company,
             ...(productBrand ? { productBrand: productBrand } : {}),
         }
         return httpClient
-            .post(`${baseURL}/v1/sellcoda/cart`, data, { params })
+            .post(`${BASE_URL_SOHEE}/v1/sellcoda/cart`, data, { params })
             .then(res => res.data)
             .catch(error => console.log(error))
     }
 
-    static getShipment(shopId: string): Promise<ShipmentEntity> {
+    static getShipment(company: string, shopId: string): Promise<ShipmentEntity> {
+        const params = { company }
         return httpClient
-            .get(`${baseURL}/v1/sellcoda/customers/dealers/${shopId}/available_shipments`)
+            .get(`${BASE_URL_SOHEE}/v1/sellcoda/customers/dealers/${shopId}/available_shipments`, { params })
             .then(res => res.data)
     }
 }