+++ src/definitions/BypassDataTest.ts
@@ -1,0 +1,1 @@
+export const bypassTelephone = ['0987654321']